+++ sdk/core/core-rest-pipeline/test/node/bearerTokenAuthenticationPolicyChallenge.spec.ts
@@ -5,13 +5,14 @@ import { assert } from "chai";
 import { AccessToken, GetTokenOptions, TokenCredential } from "@azure/core-auth";
 import {
   bearerTokenAuthenticationPolicy,
+  ChallengeCallbackOptions,
   createEmptyPipeline,
   createHttpHeaders,
   createPipelineRequest,
   HttpClient,
-  PipelineResponse
+  PipelineResponse,
+  retrieveToken
 } from "../../src";
-import { BearerTokenChallengeResult } from "../../src/policies/bearerTokenAuthenticationPolicy";
 import { TextDecoder } from "util";
 
 export interface TestChallenge {
@@ -56,7 +57,7 @@ function parseCAEChallenge(challenges: string): any[] {
     .split("Bearer ")
     .filter((x) => x)
     .map((challenge) =>
-      challenge
+      `${challenge.trim()}, `
         .split('", ')
         .filter((x) => x)
         .map((keyValue) => (([key, value]) => ({ [key]: value }))(keyValue.trim().split('="')))
@@ -64,7 +65,12 @@ function parseCAEChallenge(challenges: string): any[] {
     );
 }
 
-async function processChallenge(challenge: string): Promise<BearerTokenChallengeResult> {
+async function authenticateRequestOnChallenge(
+  challenge: string,
+  options: ChallengeCallbackOptions
+): Promise<boolean> {
+  const { scopes, setAuthorizationHeader } = options;
+
   const challenges: TestChallenge[] = parseCAEChallenge(challenge) || [];
 
   const parsedChallenge = challenges.find((x) => x.claims);
@@ -75,10 +81,19 @@ async function processChallenge(challenge: string): Promise<BearerTokenChallenge
     cachedChallenge = challenge;
   }
 
-  return {
-    scopes: [parsedChallenge.scope],
-    claims: uint8ArrayToString(decodeString(parsedChallenge.claims))
-  };
+  const accessToken = await retrieveToken({
+    ...options,
+    cachedToken: undefined,
+    scopes: parsedChallenge.scope || scopes,
+    claims: uint8ArrayToString(Buffer.from(parsedChallenge.claims, "base64"))
+  });
+
+  if (!accessToken) {
+    return false;
+  }
+
+  setAuthorizationHeader(accessToken);
+  return true;
 }
 
 class MockRefreshAzureCredential implements TokenCredential {
@@ -134,7 +149,12 @@ describe("bearerTokenAuthenticationPolicy with challenge", function() {
       scopes: "",
       credential,
       challengeCallbacks: {
-        processChallenge
+        async authenticateRequest({ cachedToken, setAuthorizationHeader }) {
+          if (cachedToken) {
+            setAuthorizationHeader(cachedToken);
+          }
+        },
+        authenticateRequestOnChallenge
       }
     });
     pipeline.addPolicy(bearerPolicy);
@@ -165,7 +185,7 @@ describe("bearerTokenAuthenticationPolicy with challenge", function() {
     assert.equal(credential.authCount, 1);
     assert.deepEqual(credential.scopesAndClaims, [
       {
-        scope: [expected.scope],
+        scope: expected.scope,
         challengeClaims: expected.challengeClaims
       }
     ]);
@@ -233,7 +253,12 @@ describe("bearerTokenAuthenticationPolicy with challenge", function() {
       scopes: "",
       credential,
       challengeCallbacks: {
-        processChallenge
+        async authenticateRequest({ cachedToken, setAuthorizationHeader }) {
+          if (cachedToken) {
+            setAuthorizationHeader(cachedToken);
+          }
+        },
+        authenticateRequestOnChallenge
       }
     });
     pipeline.addPolicy(bearerPolicy);
@@ -261,11 +286,11 @@ describe("bearerTokenAuthenticationPolicy with challenge", function() {
     assert.equal(credential.authCount, 2);
     assert.deepEqual(credential.scopesAndClaims, [
       {
-        scope: [expected[0].scope],
+        scope: expected[0].scope,
         challengeClaims: expected[0].challengeClaims
       },
       {
-        scope: [expected[1].scope],
+        scope: expected[1].scope,
         challengeClaims: expected[1].challengeClaims
       }
     ]);